unpacked/jax/output/EditableSVG/tests/requirements.txt
pytest==2.8.7
python-prctl==1.6.1
selenium==2.52.0
unittest2==1.1.0
